% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cfs_descriptives.R
\name{calculate_cfs_descriptives}
\alias{calculate_cfs_descriptives}
\title{Calculate numbers of campus file requests.}
\usage{
calculate_cfs_descriptives(cfs_input)
}
\arguments{
\item{studien_input}{A \code{data.frame} as created by \code{\link{prepare_and_extract_studien}}.}
}
\value{
A plot.
}
\description{
Calculate numbers of campus file requests.
}
\examples{
# tbd

}
