% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptives_by_year.R
\name{calculate_descriptives_by_year}
\alias{calculate_descriptives_by_year}
\title{Calculate numbers of published studies.}
\usage{
calculate_descriptives_by_year(input)
}
\arguments{
\item{input}{A \code{data.frame} as created by \code{\link{prepare_and_extract_studien}} or \code{\link{prepare_and_extract_cfs}}.}
}
\value{
A plot.
}
\description{
Calculate numbers of published studies.
}
\examples{
# tbd

}
