% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_disziplin_or_status.R
\name{calculate_disziplin_or_status}
\alias{calculate_disziplin_or_status}
\title{Calculate discipline or status frequencies.}
\usage{
calculate_disziplin_or_status(input_list, disziplin_or_status)
}
\arguments{
\item{input_list}{A \code{input_list} as created by \code{\link{prepare_and_merge_antraege}}.}

\item{disziplin_or_status}{A character vector of length 1. Should \code{"disziplin"} or \code{"status"} frequencies be calculated?.}
}
\value{
A plot.
}
\description{
Calculate data request frequencies for different scientific disciplines or status levels.
}
\examples{
# tbd

}
