% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gesamt_descriptives.R
\name{calculate_gesamt_descriptives}
\alias{calculate_gesamt_descriptives}
\title{Calculate data request frequencies.}
\usage{
calculate_gesamt_descriptives(input_list, by_var)
}
\arguments{
\item{input_list}{A \code{input_list} as created by \code{\link{prepare_and_merge_antraege}}.}

\item{by_var}{A character vector of length 1. Name of the variable.}
}
\value{
A plot.
}
\description{
Calculate data request frequencies.
}
\examples{
# tbd

}
