% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quali.R
\name{calculate_quali}
\alias{calculate_quali}
\title{Calculate qualification frequencies.}
\usage{
calculate_quali(input_list)
}
\arguments{
\item{input_list}{A \code{input_list} as created by \code{\link{prepare_and_merge_antraege}}.}
}
\value{
A plot.
}
\description{
Calculate data request frequencies for different qualification levels/phases.
}
\examples{
# tbd

}
