% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_studien_descriptives.R
\name{calculate_studien_descriptives}
\alias{calculate_studien_descriptives}
\title{Calculate data request frequencies separately for studies.}
\usage{
calculate_studien_descriptives(
  input_list,
  wellen_separat = FALSE,
  studien_regex = c("PISA", "IQB-", "IGLU", "TIMSS", "BiKS", "ELEMENT", "DESI", "StEG")
)
}
\arguments{
\item{input_list}{A \code{input_list} as created by \code{\link{prepare_and_merge_antraege}}.}

\item{wellen_separat}{A logical vector of length 1. Should different waves of longitudinal studies be counted multiple times if a
data request contains multiple ones?}

\item{studien_regex}{A character vector. Set of regular expressions selecting the studies which should be reported.}
}
\value{
A plot.
}
\description{
Calculate data request frequencies separately for studies.
}
\examples{
# tbd

}
