% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_studien_sums.R
\name{calculate_studien_sums}
\alias{calculate_studien_sums}
\title{Calculate data request frequencies.}
\usage{
calculate_studien_sums(studien_descr_jahr_all)
}
\arguments{
\item{studien_descr_jahr_all}{An object as created by \code{\link{calculate_studien_descriptives}}.}
}
\value{
A plot.
}
\description{
Calculate data request frequencies.
}
\examples{
# tbd

}
