% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_studien_uebersicht_descriptives.R
\name{calculate_studien_uebersicht_descriptives}
\alias{calculate_studien_uebersicht_descriptives}
\title{Calculate numbers of published studies.}
\usage{
calculate_studien_uebersicht_descriptives(studien_input)
}
\arguments{
\item{studien_input}{A \code{data.frame} as created by \code{\link{prepare_and_extract_studien}}.}
}
\value{
A plot.
}
\description{
Calculate numbers of published studies.
}
\examples{
# tbd

}
