% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDZReport-package.R
\docType{package}
\name{FDZReport-package}
\alias{FDZReport}
\alias{FDZReport-package}
\title{FDZReport: Create reports for FDZ data requests and data sets provided}
\description{
Create reports for FDZ data requests and data sets provided.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
