% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gesamt_descriptives.R
\name{plot_gesamt_descriptives}
\alias{plot_gesamt_descriptives}
\title{Create a line plot for absolute and cumulative frequencies.}
\usage{
plot_gesamt_descriptives(gesamt_descriptives)
}
\arguments{
\item{gesamt_descriptives}{An object as produced by \code{\link{calculate_gesamt_descriptives}}.}
}
\value{
A plot.
}
\description{
Create a line plot for absolute and cumulative frequencies.
}
\examples{
# tbd

}
