% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Create a lollipop plot.}
\usage{
plot_lollipop(dat, x, y = "Prozent", label = "label", x_axis_label)
}
\arguments{
\item{dat}{A \code{data.frame}.}

\item{x}{Character vector of length 1. Name of the column of the grouping variable.}

\item{y}{Character vector of length 1. Name of the column of the relative frequency variable.}

\item{label}{Character vector of length 1. Name of the column containing the labeling variable.}

\item{x_axis_label}{Character vector of length 1. Label for the x axis.}
}
\value{
A plot.
}
\description{
Create a lollipop plot.
}
\examples{
# tbd

}
