% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_studie_descriptives.R
\name{plot_single_studie_descriptives}
\alias{plot_single_studie_descriptives}
\title{Create a line plot for absolute and cumulative frequencies.}
\usage{
plot_single_studie_descriptives(studien_descriptives, studien_titel)
}
\arguments{
\item{studien_descriptives}{An object as produced by \code{\link{calculate_studien_descriptives}}.}

\item{studien_titel}{Character vector of length 1. Name of the study which is being plotted.}
}
\value{
A plot.
}
\description{
Create a line plot for absolute and cumulative frequencies.
}
\examples{
# tbd

}
