% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_studien_descriptives.R
\name{plot_studien_descriptives}
\alias{plot_studien_descriptives}
\title{Create a line plot for absolute and cumulative frequencies.}
\usage{
plot_studien_descriptives(studien_descriptives)
}
\arguments{
\item{studien_descriptives}{An object as produced by \code{\link{calculate_studien_descriptives}}.}
}
\value{
A plot.
}
\description{
Create a line plot for absolute and cumulative frequencies.
}
\examples{
# tbd

}
