% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_studien_ranking.R
\name{plot_studien_ranking}
\alias{plot_studien_ranking}
\title{Create a bar plot with rank ordered cumulative frequencies.}
\usage{
plot_studien_ranking(studien_sums)
}
\arguments{
\item{studien_sums}{An object as produced by \code{\link{calculate_studien_sums}}.}
}
\value{
A plot.
}
\description{
Create a bar plot with rank ordered cumulative frequencies.
}
\examples{
# tbd

}
