% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_extract_CFs.R
\name{prepare_and_extract_cfs}
\alias{prepare_and_extract_cfs}
\title{Prepare and extract data of campus files data requests.}
\usage{
prepare_and_extract_cfs(cfs_uebersicht)
}
\arguments{
\item{cfs_uebersicht}{A \code{.xlsm} file including all data requests for campus files.}
}
\value{
A plot.
}
\description{
Prepare and extract data of campus files data requests.
}
\examples{
# tbd

}
