% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_extract_studien.R
\name{prepare_and_extract_studien}
\alias{prepare_and_extract_studien}
\title{Prepare and merge data of data requests.}
\usage{
prepare_and_extract_studien(studien_uebersicht)
}
\arguments{
\item{studien_uebersicht}{A \code{.xlsm} file including all archieved studies at the FDZ, as produced by \code{StudienDB}.}
}
\value{
A plot.
}
\description{
Load, prepare and merge data of data requests.
}
\examples{
# tbd

}
