% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_merge_antraege.R
\name{prepare_and_merge_antraege}
\alias{prepare_and_merge_antraege}
\title{Prepare and merge data of data requests.}
\usage{
prepare_and_merge_antraege(antraege_aktiv, antraege_archiv)
}
\arguments{
\item{antraege_aktiv}{A \code{.xlsm} file including all active data requests, as produced by \code{FAMe}.}

\item{antraege_archiv}{A \code{.xlsm} file including all archived data requests, as produced by \code{FAMe}.}
}
\value{
A plot.
}
\description{
Load, prepare and merge data of data requests.
}
\examples{
# tbd

}
