% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gesamt_descriptives.R
\name{plot_gesamt_descriptives}
\alias{plot_gesamt_descriptives}
\title{Create a line plot for absolute and cumulative frequencies.}
\usage{
plot_gesamt_descriptives(
  gesamt_descriptives,
  ylab,
  ylim = c(NA_integer_, NA_integer_)
)
}
\arguments{
\item{gesamt_descriptives}{An object as produced by \code{\link{calculate_gesamt_descriptives}}.}

\item{ylab}{The title of the y axis.}

\item{ylim}{The limits of the y axis (pre default calculated based on the variable range).}
}
\value{
A plot.
}
\description{
Create a line plot for absolute and cumulative frequencies.
}
\examples{
# tbd

}
